requirements.txt
pandas==2.2.2
numpy==2.0.1
yfinance==1.0
alpha_vantage==3.0.0
qdrant-client==1.16.0
langgraph==1.0.0
langchain-core==1.0.0
langchain-openai==1.0.0
pydantic==2.11.7
python-dotenv==1.1.0
fastmcp==2.14.1
click==8.1.7
openai==2.12.0